chore(rename): Container is now a plain JS file
